Imaging model with stationary atoms
Importing of Isat calibration data from Sept. 17 2013 is in importFiles.
isat.py takes the imaging calibration data and deduces the average OD in
a small box for each file
ImagingModel3 simulates absorption of light by atoms for varying light
intensites and imaging times. It takes into account recoil detuning, but
does not allow the atoms to move.
compare.py compares the simulated od with the calculated od from the data,
and finds the standard deviation of the od for different values of isat
in counts - ie. finds the value of isat that gives the most consistent results